requirements.txt
fastapi
uvicorn
openai
numpy
python-dotenv
